Update README.md
# ipl_batsman_analysis
This is the script and datasets for analysing and comparing 3 batsmen in IPL (2008 to 2019). If you are a cricket enthusiast, this will give you lots of insights into the weakness of batsmen as well. 

Multiple charts and breaks by weekday, venue, opponent etc is also provided. You just need to update the variables (bat_player_1, bat_player_2 & bat_player_3) as any player name from deliveries.csv.

A folder named as Analysis_Results would be created and the analysis result would be updated in it as a 5 page pdf file. Individual pages are provided as png file. 2 pages are comparison chart and 3 are individual player's IPL analysis.